requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
ultralytics>=8.0.0
facenet-pytorch>=2.5.3
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
huggingface-hub>=0.20.0
